+++ mobile/app/components/atoms/Image.js
@@ -1,5 +1,11 @@
 import React from 'react';
-import {Image} from 'react-native';
+import {Image, StyleSheet} from 'react-native';
 export default (props) => {
-    return <Image {...props} />
+    return <Image {...props} style={styles.image} resizeMode="contain" />
 }
+const styles =StyleSheet.create({
+    image:{
+        width: "100%",
+        height: '100%',
+    }
+})
